Make download link into button + open in new tab
The download link sitting next to two buttons on the page looks kind of out of place. This patch allows it to retain the functionality of a link, but looks more consistent with the other buttons on the page. Also, since it no longer looks like a link, the user might not know to right click->open in new tab or middle click in order to not replace the current window and potentially lose work, so the download will open in a new tab by default.

(This is kind of a personal stylistic preference change, feel free to close if you don't like it.)