Make play/stop buttons toggle
Currently, the play/stop buttons are both present and clickable whether
audio is playing or not. It's not possible to visually tell whether
playback is occuring, which is frustrating when your audio contains long
periods of silence.

This change makes it so only one button is visible at a time. When
playback starts, the play button gets replaced by the stop button, so
the user doesn't need to move their mouse to stop playback. This also
provides a clear visual indication for whether audio is playing.