Move download link out of form
Having the download link on the same line as the play/stop button looks
unorganized and weird. This change moves it out of the <form></form>
entirely and wraps it in its own <p></p>.

The link's inner text is also changed to better reflect its function,
and the target attribute is set to _blank to avoid replacing the current
tab and losing work. This change was previously okayed:
https://github.com/calzoneman/aeiou/pull/10#issuecomment-1133809367